0.8.3665

+++ amulet_engine/lib/classes/amulet_game.dart
@@ -434,7 +434,7 @@ class AmuletGame extends IsometricGame<AmuletPlayer> {
     required Collider target,
     required double damage,
     required DamageType damageType,
-    required int ailmentDuration,
+    required double ailmentDuration,
     required double ailmentDamage,
     double? angle,
     double force = 0,
